feat/#123 : Constant refactoring
